fix(resource-recorder)!: disable service id endpoint (#1644)

+++ resource-recorder/src/lib.rs
@@ -1,7 +1,10 @@
 use async_trait::async_trait;
 use dal::{Dal, DalError, Resource};
 use prost_types::TimestampError;
-use shuttle_common::{backends::auth::VerifyClaim, claims::Scope};
+use shuttle_common::{
+    backends::{auth::VerifyClaim, client::gateway, ClaimExt},
+    claims::{Claim, Scope},
+};
 use shuttle_proto::resource_recorder::{
     self, resource_recorder_server::ResourceRecorder, ProjectResourcesRequest, RecordRequest,
     ResourceIds, ResourceResponse, ResourcesResponse, ResultResponse, ServiceResourcesRequest,
@@ -42,14 +45,18 @@ impl From<String> for Error {
 
 pub struct Service<D> {
     dal: D,
+    gateway_client: gateway::Client,
 }
 
 impl<D> Service<D>
 where
     D: Dal + Send + Sync + 'static,
 {
-    pub fn new(dal: D) -> Self {
-        Self { dal }
+    pub fn new(dal: D, gateway_client: gateway::Client) -> Self {
+        Self {
+            dal,
+            gateway_client,
+        }
     }
 
     /// Record the addition of a new resource
@@ -87,18 +94,6 @@ where
         Ok(resources.into_iter().map(Into::into).collect())
     }
 
-    /// Get the resources that belong to a service
-    async fn service_resources(
-        &self,
-        service_id: String,
-    ) -> Result<Vec<resource_recorder::Resource>, Error> {
-        tracing::info!("fetching resources for service");
-
-        let resources = self.dal.get_service_resources(service_id.parse()?).await?;
-
-        Ok(resources.into_iter().map(Into::into).collect())
-    }
-
     /// Get a resource
     async fn get_resource(
         &self,
@@ -120,6 +115,21 @@ where
 
         Ok(())
     }
+
+    async fn verify_ownership(&self, claim: &Claim, project_id: &str) -> Result<(), Status> {
+        if !claim.is_admin()
+            && !claim.is_deployer()
+            && !claim
+                .owns_project_id(&self.gateway_client, project_id)
+                .await
+                .map_err(|_| Status::internal("could not verify project ownership"))?
+        {
+            let status = Status::permission_denied("the request lacks the authorizations");
+            error!(error = &status as &dyn std::error::Error);
+            return Err(status);
+        }
+        Ok(())
+    }
 }
 
 #[async_trait]
@@ -133,8 +143,9 @@ where
         request: Request<RecordRequest>,
     ) -> Result<Response<ResultResponse>, Status> {
         request.verify(Scope::ResourcesWrite)?;
-
+        let claim = request.get_claim()?;
         let request = request.into_inner();
+        self.verify_ownership(&claim, &request.project_id).await?;
 
         let result = match self.add(request).await {
             Ok(()) => ResultResponse {
@@ -156,8 +167,10 @@ where
         request: Request<ProjectResourcesRequest>,
     ) -> Result<Response<ResourcesResponse>, Status> {
         request.verify(Scope::Resources)?;
-
+        let claim = request.get_claim()?;
         let request = request.into_inner();
+        self.verify_ownership(&claim, &request.project_id).await?;
+
         let result = match self.project_resources(request.project_id).await {
             Ok(resources) => ResourcesResponse {
                 success: true,
@@ -177,25 +190,11 @@ where
     #[tracing::instrument(skip(self))]
     async fn get_service_resources(
         &self,
-        request: Request<ServiceResourcesRequest>,
+        _request: Request<ServiceResourcesRequest>,
     ) -> Result<Response<ResourcesResponse>, Status> {
-        request.verify(Scope::Resources)?;
-
-        let request = request.into_inner();
-        let result = match self.service_resources(request.service_id).await {
-            Ok(resources) => ResourcesResponse {
-                success: true,
-                message: Default::default(),
-                resources,
-            },
-            Err(e) => ResourcesResponse {
-                success: false,
-                message: e.to_string(),
-                resources: Vec::new(),
-            },
-        };
-
-        Ok(Response::new(result))
+        Err(Status::not_found(
+            "This resource endpoint is discontinued. Please restart your project.",
+        ))
     }
 
     #[tracing::instrument(skip(self))]
@@ -204,8 +203,10 @@ where
         request: tonic::Request<ResourceIds>,
     ) -> Result<Response<ResourceResponse>, Status> {
         request.verify(Scope::Resources)?;
-
+        let claim = request.get_claim()?;
         let request = request.into_inner();
+        self.verify_ownership(&claim, &request.project_id).await?;
+
         let result = match self.get_resource(request).await {
             Ok(resource) => ResourceResponse {
                 success: true,
@@ -215,7 +216,7 @@ where
             Err(e) => ResourceResponse {
                 success: false,
                 message: e.to_string(),
-                resource: Default::default(),
+                resource: None,
             },
         };
 
@@ -228,8 +229,10 @@ where
         request: tonic::Request<ResourceIds>,
     ) -> Result<Response<ResultResponse>, Status> {
         request.verify(Scope::ResourcesWrite)?;
-
+        let claim = request.get_claim()?;
         let request = request.into_inner();
+        self.verify_ownership(&claim, &request.project_id).await?;
+
         let result = match self.delete_resource(request).await {
             Ok(()) => ResultResponse {
                 success: true,
